Move tests related to toDot() method to a separate class 'DotDisplayTest'.
